# tests/openapi_baseline.json
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "REST API for ListSync media synchronization dashboard",
    "title": "ListSync Web UI API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/collections": {
      "get": {
        "description": "Get all collections with pagination, search, and sorting",
        "operationId": "get_collections_api_collections_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Search collections by franchise name",
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "default": "",
              "description": "Search collections by franchise name",
              "title": "Search",
              "type": "string"
            }
          },
          {
            "description": "Sort by popularity, rating, movie_count, name, or total_votes",
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "total_votes",
              "description": "Sort by popularity, rating, movie_count, name, or total_votes",
              "pattern": "^(popularity|rating|movie_count|name|total_votes)$",
              "title": "Sort",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Collections"
      }
    },
...
    "/api/collections/random": {
      "get": {
        "description": "Get random collections from the full database (only collections with 3+ movies) - cached for performance",
        "operationId": "get_random_collections_api_collections_random_get",
        "parameters": [
          {
            "description": "Number of random collections to return",
            "in": "query",
            "name": "count",
            "required": false,
            "schema": {
              "default": 5,
              "description": "Number of random collections to return",
              "maximum": 20,
              "minimum": 1,
              "title": "Count",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Random Collections"
      }
    },
...
    "/api/collections/{franchise_name}": {
      "get": {
        "description": "Get specific collection details by franchise name",
        "operationId": "get_collection_details_api_collections__franchise_name__get",
        "parameters": [
          {
            "in": "path",
            "name": "franchise_name",
            "required": true,
            "schema": {
              "title": "Franchise Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Collection Details"
      }
    },
    "/api/collections/{franchise_name}/movies": {
      "get": {
        "description": "Get movies in a collection with full details",
        "operationId": "get_collection_movies_api_collections__franchise_name__movies_get",
        "parameters": [
          {
            "in": "path",
            "name": "franchise_name",
            "required": true,
            "schema": {
              "title": "Franchise Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Collection Movies"
      }
    },
    "/api/collections/{franchise_name}/poster": {
      "get": {
        "description": "Get poster URL for collection (uses most voted movie's poster from Trakt)",
        "operationId": "get_collection_poster_api_collections__franchise_name__poster_get",
        "parameters": [
          {
            "in": "path",
            "name": "franchise_name",
            "required": true,
            "schema": {
              "title": "Franchise Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Collection Poster"
      }
    },
    "/api/collections/{franchise_name}/sync": {
      "post": {
        "description": "Sync a collection to Seerr using a terminable subprocess",
        "operationId": "sync_collection_api_collections__franchise_name__sync_post",
        "parameters": [
          {
            "in": "path",
            "name": "franchise_name",
            "required": true,
            "schema": {
              "title": "Franchise Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sync Collection"
      }
    },
...
  }
}